feat: 🎸 (dialysis) new ui
